test: finish, running package 이동
