+++ src/test/java/bowling/domain/state/finish/SpareTest.java
@@ -1,6 +1,7 @@
-package bowling.domain.state;
+package bowling.domain.state.finish;
 
 import bowling.domain.score.Score;
+import bowling.domain.state.State;
 import bowling.domain.state.finish.Spare;
 import bowling.exception.ImpossiblePitchException;
 import bowling.domain.Pins;
