+++ src/test/java/bowling/domain/state/finish/StrikeTest.java
@@ -1,7 +1,8 @@
-package bowling.domain.state;
+package bowling.domain.state.finish;
 
 import bowling.domain.score.Score;
-import bowling.domain.state.finish.Miss;
+import bowling.domain.state.State;
+import bowling.domain.state.finish.Strike;
 import bowling.exception.ImpossiblePitchException;
 import bowling.domain.Pins;
 import org.junit.jupiter.api.DisplayName;
@@ -11,44 +12,42 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-class MissTest {
+class StrikeTest {
 
-    private final State miss = Miss.of(Pins.create(5), Pins.create(4));
+    private final State strike = Strike.create();
 
     @Test
     @DisplayName("생성을 확인한다")
     void create() {
-        assertThat(miss).isInstanceOf(Miss.class);
+        assertThat(strike).isInstanceOf(Strike.class);
     }
 
     @Test
-    @DisplayName("심볼은 두 투구의 숫자를 반환한다")
+    @DisplayName("심볼을 확인한다")
     void getSymbol() {
-        assertThat(miss.getSymbol()).isEqualTo("5|4");
+        assertThat(strike.getSymbol()).isEqualTo("X");
     }
 
     @Test
     @DisplayName("프레임을 종료한 상태이다")
-    void name() {
-        assertThat(miss.isFrameEnd()).isTrue();
+    void isFrameStateEnd() {
+        assertThat(strike.isFrameEnd()).isTrue();
     }
 
     @Test
-    @DisplayName("더이상 투구를 할 수 없다")
-    void dfa() {
-        assertThatThrownBy(() -> miss.pitch(Pins.create(3))).isInstanceOf(ImpossiblePitchException.class);
+    @DisplayName("투구를 할 수 없다")
+    void doNotPitch() {
+        assertThatThrownBy(() -> strike.pitch(Pins.create(3))).isInstanceOf(ImpossiblePitchException.class);
     }
 
     @Test
-    @DisplayName("거터를 포함할 경우, (-) 심볼을 추가한다")
-    void getSymbolIncludedGutter() {
-        //given, when
-        Pins firstPins = Pins.create(0);
-        Pins secondPins = Pins.create(4);
-        State missIncludedGutter = Miss.of(firstPins, secondPins);
-
-        //then
-        assertThat(missIncludedGutter.getSymbol()).isEqualTo("-|4");
+    @DisplayName("스코어 반환을 확인한다")
+    void score() {
+        Score score = strike.score();
+        assertAll(
+                () -> assertThat(score.finishCalculation()).isFalse(),
+                () -> assertThat(score.score()).isEqualTo(10)
+        );
     }
 
     @Test
@@ -57,7 +56,7 @@ class MissTest {
         Score beforeScore = Score.miss(Pins.create(1));
         int expectedScore = beforeScore.score();
 
-        Score score = miss.calculateScore(beforeScore);
+        Score score = strike.calculateScore(beforeScore);
 
         assertAll(
                 () -> assertThat(score.finishCalculation()).isTrue(),
@@ -69,9 +68,9 @@ class MissTest {
     @DisplayName("이전 결과가 Spare 이면 이전 스코어와 합산을 확인한다")
     void calculateScoreWhenBeforeStateIsSpare() {
         Score beforeScore = Score.spare();
-        int expectedScore = beforeScore.score() + Pins.create(5).count();
+        int expectedScore = beforeScore.score() + strike.score().score();
 
-        Score score = miss.calculateScore(beforeScore);
+        Score score = strike.calculateScore(beforeScore);
 
         assertAll(
                 () -> assertThat(score.finishCalculation()).isTrue(),
@@ -83,12 +82,12 @@ class MissTest {
     @DisplayName("이전 결과가 Strike 이면 이전 스코어와 합산을 확인한다")
     void calculateScoreWhenBeforeStateIsStrike() {
         Score beforeScore = Score.strike();
-        int expectedScore = beforeScore.score() + miss.score().score();
+        int expectedScore = beforeScore.score() + strike.score().score();
 
-        Score score = miss.calculateScore(beforeScore);
+        Score score = strike.calculateScore(beforeScore);
 
         assertAll(
-                () -> assertThat(score.finishCalculation()).isTrue(),
+                () -> assertThat(score.finishCalculation()).isFalse(),
                 () -> assertThat(score.score()).isEqualTo(expectedScore)
         );
     }
