+++ src/test/java/bowling/domain/state/running/ReadyTest.java
@@ -1,6 +1,7 @@
-package bowling.domain.state;
+package bowling.domain.state.running;
 
 import bowling.domain.Pins;
+import bowling.domain.state.State;
 import bowling.domain.state.finish.Strike;
 import bowling.domain.state.running.FirstBowl;
 import bowling.domain.state.running.Ready;
